ML/requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests
transformers
langchain-huggingface
langchain-ollama